DIP - Diplomová práca/Dataset generating/requirements.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
future==0.18.2
gast==0.2.2
ghostscript==0.6
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.13.2
grpcio==1.25.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
imread==0.7.1
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-tuner==1.0.1
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mysqlclient==1.4.5
netron==3.9.8
networkx==2.4
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandas==1.0.1
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pydot==1.4.1
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.2.0
six==1.13.0
tabulate==0.8.7
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminaltables==3.1.0
tqdm==4.43.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
